fix: resolve bschwartz PR review comments
